+++ tests/lint/freeandnil-on-interface.pas
@@ -1,0 +1,15 @@
+unit FanIntf;
+
+interface
+
+uses System.SysUtils;
+
+implementation
+
+procedure P(Intf: IInterface; Obj: TObject);
+begin
+  FreeAndNil(Intf);
+  FreeAndNil(Obj);
+end;
+
+end.